# client-sdks/manager/rust/openapi-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alien Manager API",
    "description": "Control plane for Alien applications. Manages deployments, releases, commands, and telemetry.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/deployments/{id}": {
      "get": {
        "tags": [
          "deployments"
        ],
        "operationId": "get_deployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Deployment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "deployments"
        ],
        "operationId": "delete_deployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Deployment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "Force delete without running cleanup (immediately removes record)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deployment deletion enqueued"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeploymentGroupMinimal": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "DeploymentResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "platform",
          "status",
          "deploymentGroupId",
          "retryRequested",
          "createdAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "currentReleaseId": {
            "type": "string",
            "nullable": true
          },
          "deploymentGroup": {
            "$ref": "#/components/schemas/DeploymentGroupMinimal",
            "nullable": true
          },
          "deploymentGroupId": {
            "type": "string"
          },
          "desiredReleaseId": {
            "type": "string",
            "nullable": true
          },
          "environmentInfo": {},
          "error": {},
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "retryRequested": {
            "type": "boolean"
          },
          "runtimeMetadata": {},
          "stackSettings": {},
          "stackState": {},
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Platform": {
        "type": "string",
        "description": "Represents the target cloud platform.",
        "enum": [
          "aws",
          "gcp",
          "azure",
          "kubernetes",
          "local",
          "test"
        ]
      },
...
    }
  },
...
}